+++ CHANGELOG.md
@@ -1,11 +1,13 @@
 # Changelog
 
-## Version 0.1.2 (?)
+## Version 0.1.2 (2014-07-22)
 
 * Correcting small bug that occurred when setSelectionRange() was called on 
   an element that does not support it.
 * Fix Pointer Event bug: Too many event listeners in move, end, cancel.
 * Fix for Bug #1 - Not working in Windows 8.1 IE11
+* Adding a `cancelled` flag to `DraggableEvent` to indicate if a drag ended 
+  because of a cancelling operation like `esc` key, etc.
 
 
 ## Version 0.1.1 (2014-07-21)